One line change, big question: src/org/openintents/safe/PassEdit.java
@@ -143,6 +143,7 @@
   @Override
   protected void onPause() {
   	     super.onPause();
+		saveState();
		dbHelper.close();
		dbHelper = null;
	}

We now save the state on Pause. This allows the user's work to get
saved if they hit hte "Home" key, or if the application logs out
because of a timeout or because of the user selecting "log out" from
the notification area.

It also will save the data if the user hits the "back" key.

Is this OK?