Refactored the main intents in order to use permissions.
Now, to use any of the intents: get/set encrypt/decrypt, you will need
to declare use of a permission as follows:
    <uses-permission android:name="org.openintents.safe.ACCESS_INTENTS" />

I've made that change to TestSafe and Notepad.

This required refactoring the FrontDoor.  Basically, I moved _all_ the
functionality of FrontDoor to a new class: IntentHandler, and changed
the name of that class. Now FrontDoor basically only:

    if (action == null || action.equals(Intent.ACTION_MAIN)){
    	  Intent i = new Intent(getApplicationContext(),
          IntentHandler.class);
          startActivity(i);
    } 

Now FrontDoor only handles those actions as above, and IntentHandler
handles all the others, and includes the permission:

   <activity class=".IntentHandler" android:name="IntentHandler" android:label="@string/app_name"
              android:permission="org.openintents.safe.ACCESS_INTENTS">

Please see the OI Safe Manifest file for details.

This was required because the launcher would need to use that
permission if it protected ACTION_MAIN!  So now the launcher is
allowed to run FrontDoor (which requires no permission) and FrontDoor
runs IntentHandler.  Basically this goes around the permission in the
case of action MAIN.