OI File Manager: Apply patch by Dominik Pretzsch: Copy of file - internationalization.
If a file exists having the same name, it'll split the file's name and extension. 
After that fileName is passed to getString() without extension to build the localized "Copy of"-String, which is then concatenated with correct extension. 

For files without extension or linux style hidden files (e.g '.foo') the file name is not split. 

Fixes issue 307.