SensorSimulator: apply patch by Prithvi Raj: adds Record and Playback buttons.
Data are written to SensorData.txt.
There are some cases in which playback is not the same data 
in the file due to the way the sliders are set. 
The playback happens at the rate that is specified in the GUI at the 
time of the playback, and not which was set during recording.