OI Shopping List: Fix crash launching on ICS. Fixes issue 403.
Instead of using a hard-coded check for HoneyComb to check whether
the code thinks we are running in tablet mode, check for the existence
of the R.id.list resource, whose absence in the non-tablet resources was
causing us to crash.

With this change, Honeycomb is still running in fragment / tablet mode. ICS no
longer crashes but it doesn't seem to run in tablet mode either. I tried
two different emulator configurations, one with phone-like HVGA dimensions
and the other at 1280x800, and neither one chose the tablet resources.

Further study is required to figure out how to make ICS choose the
tablet-style resources when running on larger devices. Assuming that
might be done by adding more flavors of resources, this code should not
need to be changed when it happens.

http://developer.android.com/guide/topics/resources/providing-resources.html#Compatibility