OI Shopping List: Quick Edit Mode.
New preference is off by default, and disabled on Cupcake and
Donut. When enabled, Quantity and Priority use a PopupMenu instead
of launching the Edit Items dialog. Popup menu contains numbers
1 through 5 and "Other Priority..." or "Other Quantity...".

The popup menu is implemented using a wrapper class which choosed
android.widget.PopupMenu on Honeycomb or ICS, and on Froyo or
Gingerbread it uses a built-in PopupMenu which is backported
from ICS source.