OI Notepad: GCI task by Daniel Huang: apply marquee setting immediately.
Google Code-in task:
http://www.google-melange.com/gci/task/view/google/gci2011/7134322

Update issue 405
The first problem mentioned for Android 4.0 is solved:

The issue is mentioned in the page of issue #405:
"If you switch to settings, 
change the marquee setting, and press back, the change is not visible immediately. One has to leave the application and start OI Notepad again."
 
Steps to reproduce the problem:

Open notepad application
- create one note with longer title
- menu-> settings -> enable/or disable marquee setting 
- Hit Back, view switch to notepad again from setting
 - look at the title of the note, marquee setting is not take effect, 
- but if you close notepad application and open it again, marquee is observed for note title.

Root cause:
Marquee handling for title is in constructor, but the constructor is not getting called when the view is changed back from Settings to the notepad application. This causes the 
 marquee setting to not take effect after switching back from Settings to the notepad view.

Solution:
Move the marquee handling constructor to hasFocus(), which is called whenever the notepad view gets focus.
 
Verification:

1. Enable/disable marquee from Settings, go back to the notepad view, marquee enable/disable should be observed on title
2. Enable/disable marquee from Settings, close notepad application, open notepad application again, marquee enable/disable should be observed on title